Merge ground control points of 2 datasets for further comparison

This function takes two outputs of prepare_gcps() and merges them for comparison.

Usage

merge_GCPs(
  gcp_data_test,
  gcp_data_ref,
  export = F,
  filename = NULL,
  output_path = NULL
)

Arguments

gcp_data_test

object of class spatial points as generated by bfnpALSprocessor::prepare_gcps()). This is the dataset the accurracy should be checked for

gcp_data_ref

object of class spatial points as generated by bfnpALSprocessor::prepare_gcps()). This is the reference dataset (i.e. the ALS 2019-2020 dataset)

export

logical of length 1. Should the result be exported to a geopackage file? Defaults to FALSE.

filename

character, optional, name of the resulting geopackage file

output_path

character, optional, path to folder where the newly generated gpkg file will be exported to

Value

sf points

Examples